--- a/vllm/entrypoints/score_utils.py
+++ b/vllm/entrypoints/score_utils.py
@@ -1,15 +1,17 @@
 # SPDX-License-Identifier: Apache-2.0
-from typing import Union
+from typing import TYPE_CHECKING, Union
 
 from torch.nn import CosineSimilarity
 
 from vllm.outputs import PoolingRequestOutput
-from vllm.transformers_utils.tokenizer import (PreTrainedTokenizer,
-                                               PreTrainedTokenizerFast)
+
+if TYPE_CHECKING:
+    from vllm.transformers_utils.tokenizer import (PreTrainedTokenizer,
+                                                   PreTrainedTokenizerFast)
 
 
 def _cosine_similarity(
-    tokenizer: Union[PreTrainedTokenizer, PreTrainedTokenizerFast],
+    tokenizer: Union["PreTrainedTokenizer", "PreTrainedTokenizerFast"],
     embed_1: list[PoolingRequestOutput],
     embed_2: list[PoolingRequestOutput],
 ) -> list[PoolingRequestOutput]: