FIX https://github.com/vllm-project/vllm/issues/14924