@Chen-0210 2025-03-25
@aarnphm Thanks for your review. I’ve completed round 2.