@davidxia 2025-04-22
@Chen-0210 I made some fixes to this PR in https://github.com/Chen-0210/vllm/pull/1 to get `vllm --version` working. I then ran benchmarks for both with [hyperfine](https://github.com/sharkdp/hyperfine) on a Linux x86 VM with 8 CPUs, 32GB memory, running Ubuntu 24.04 (noble) with Python 3.12.3.

## `python -c 'import vllm'`

**Faster**

### before (main branch commit 5536b30a4c7877d75758d21bdaf39b3a59aa2dc2)

```shell
$ hyperfine 'python -c "import vllm"' --warmup 3 --runs 20 --export-markdown out.md
Benchmark 1: python -c "import vllm"
  Time (mean ± σ):      7.316 s ±  0.104 s    [User: 7.254 s, System: 0.857 s]
  Range (min … max):    7.161 s …  7.511 s    20 runs
```
 
| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `python -c "import vllm"` | 7.316 ± 0.104 | 7.161 | 7.511 | 1.00 |

### after (my PR commit 31074c25d4c0fb52e48eec5c359eb7a9703ad06f)

```shell
$ hyperfine 'python -c "import vllm"' --warmup 3 --runs 20 --export-markdown out.md
Benchmark 1: python -c "import vllm"
  Time (mean ± σ):      4.955 s ±  0.133 s    [User: 5.163 s, System: 0.609 s]
  Range (min … max):    4.755 s …  5.242 s    20 runs
```

| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `python -c "import vllm"` | 4.955 ± 0.133 | 4.755 | 5.242 | 1.00 |


## `vllm --version`

**Not much improvement, but like you said in PR description, we can optimize in separate PR.**

### before (main branch commit 5536b30a4c7877d75758d21bdaf39b3a59aa2dc2)

```shell
$ hyperfine 'vllm --version' --warmup 3 --runs 20 --export-markdown out.md
Benchmark 1: vllm --version
  Time (mean ± σ):     10.033 s ±  0.203 s    [User: 9.970 s, System: 0.879 s]
  Range (min … max):    9.745 s … 10.577 s    20 runs
```
 
| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `vllm --version` | 10.033 ± 0.203 | 9.745 | 10.577 | 1.00 |

### after (my PR commit 31074c25d4c0fb52e48eec5c359eb7a9703ad06f)

```shell
$ hyperfine 'vllm --version' --warmup 3 --runs 20 --export-markdown out.md
Benchmark 1: vllm --version
  Time (mean ± σ):      9.248 s ±  0.203 s    [User: 9.104 s, System: 0.956 s]
  Range (min … max):    8.900 s …  9.736 s    20 runs
```
 
| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `vllm --version` | 9.248 ± 0.203 | 8.900 | 9.736 | 1.00 |