@davidxia 2025-04-23
I reran the benchmarks with the same setup. Compare with the before here.

### commit 36a0bc1e5bf5621ab42accca24ee454d9e18583e

## `python -c 'import vllm'`

**52% faster!**

```shell
$ hyperfine 'python -c "import vllm"' --warmup 3 --runs 20 --export-markdown out.md
Benchmark 1: python -c "import vllm"
  Time (mean ± σ):      3.504 s ±  0.119 s    [User: 3.947 s, System: 0.366 s]
  Range (min … max):    3.306 s …  3.891 s    20 runs
```
 
| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `python -c "import vllm"` | 3.504 ± 0.119 | 3.306 | 3.891 | 1.00 |


## `vllm --version`

### commit 36a0bc1e5bf5621ab42accca24ee454d9e18583e

```shell
$ hyperfine 'vllm --version' --warmup 3 --runs 20 --export-markdown out.md
Benchmark 1: vllm --version
  Time (mean ± σ):      9.853 s ±  0.329 s    [User: 9.643 s, System: 0.973 s]
  Range (min … max):    9.515 s … 10.599 s    20 runs
```
 
| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `vllm --version` | 9.853 ± 0.329 | 9.515 | 10.599 | 1.00 |