@reidliu41 2025-05-10
lol, Just tested the latest code, since currently, it will load more and more things..

```
vllm --help
INFO 05-10 16:58:03 [__init__.py:248] Automatically detected platform cuda.
INFO 05-10 16:58:10 [config.py:760] This model supports multiple tasks: {'reward', 'score', 'generate', 'embed', 'classify'}. Defaulting to 'generate'.
WARNING 05-10 16:58:10 [config.py:2019] max_model_len was is not set. Defaulting to arbitrary value of 8192.
WARNING 05-10 16:58:10 [config.py:2025] max_num_seqs was is not set. Defaulting to arbitrary value of 128.
INFO 05-10 16:58:11 [config.py:760] This model supports multiple tasks: {'reward', 'score', 'generate', 'embed', 'classify'}. Defaulting to 'generate'.
INFO 05-10 16:58:12 [config.py:760] This model supports multiple tasks: {'reward', 'score', 'generate', 'embed', 'classify'}. Defaulting to 'generate'.
INFO 05-10 16:58:13 [config.py:760] This model supports multiple tasks: {'reward', 'score', 'generate', 'embed', 'classify'}. Defaulting to 'generate'.
usage: vllm [-h] [-v] {chat,complete,serve,bench,collect-env} ...

vLLM CLI

positional arguments:
  {chat,complete,serve,bench,collect-env}
    chat                Generate chat completions via the running API server.
    complete            Generate text completions based on the given prompt via the running API server.
    serve               Start the vLLM OpenAI Compatible API server.
    bench               vLLM bench subcommand.
    collect-env         Start collecting environment information.

options:
  -h, --help            show this help message and exit
  -v, --version         show program's version number and exit


$ vllm --version
INFO 05-10 14:25:32 [__init__.py:248] Automatically detected platform cuda.
ERROR 05-10 14:30:34 [config.py:108] Error retrieving file list: (MaxRetryError("HTTPSConnectionPool(host='huggingface.co', port=443): Max retries exceeded with url: /api/models/facebook/opt-125m/tree/main?recursive=True&expand=False (Caused by ConnectTimeoutError(<urllib3.connection.HTTPSConnection object at 0x7f9b59e508f0>, 'Connection to huggingface.co timed out. (connect timeout=None)'))"), '(Request ID: b598ef4d-db37-4a4a-aa4b-6e0e4fd561a8)'), retrying 1 of 2
```