@davidxia 2025-05-12
> I would suggest to try at smaller scale iteratively, let's start at multimodal directory?

https://github.com/vllm-project/vllm/pull/18031